upstream: b=master,r=f602849a2d013a4566861ff1a5622dbc00f12a55,t=2018-12-07-1256-50858

+++ components/nexus-repository/src/main/java/org/sonatype/nexus/repository/rest/internal/resources/AssetsResource.java
@@ -146,7 +146,9 @@ public class AssetsResource
     DetachedEntityId entityId = new DetachedEntityId(repositoryItemIDXO.getId());
     Asset asset = getAsset(id, repository, entityId);
 
-    maintenanceService.deleteAsset(repository, asset);
+    if (repository != null && asset != null) {
+      maintenanceService.deleteAsset(repository, asset);
+    }
   }
 
   private Asset getAsset(final String id, final Repository repository, final DetachedEntityId entityId)
